Updated to use WhatsApp templating
I've updated the code, as I no longer use a Twilio sandbox. I've
now registered to use WhatsApp as a Business and use a WhatsApp template to send notifications.

It's unclear to me why the TemplateParameters param is required by the
Twilio API. The call works even if the parameters are random numbers or are
completely blank, regardless of the values in those parameters in the
body string. So long as the body matches the WhatsApp template to some extent.